Remove client.login dependency from minionUtils
