+++ src/lib/util/minionUtils.ts
@@ -1,63 +1,20 @@
-import type { activity_type_enum, User } from '@prisma/client';
-import { calcPercentOfNum, calcWhatPercent, Time } from 'e';
+import type { User } from '@prisma/client';
 import { KlasaUser } from 'klasa';
 import { Bank } from 'oldschooljs';
 import { ItemBank } from 'oldschooljs/dist/meta/types';
 
 import { getUserGear } from '../../mahoji/mahojiSettings';
-import { PerkTier } from '../constants';
 import { allPetIDs } from '../data/CollectionsExport';
 import { getSimilarItems } from '../data/similarItems';
 import { UserSettings } from '../settings/types/UserSettings';
 import { SkillsEnum } from '../skilling/types';
 import { convertXPtoLVL } from '../util';
-import getUsersPerkTier, { patronMaxTripCalc } from './getUsersPerkTier';
 import resolveItems from './resolveItems';
 
 export function skillLevel(user: KlasaUser | User, skill: SkillsEnum) {
 	const xp =
 		user instanceof KlasaUser ? (user.settings.get(`skills.${skill}`) as number) : Number(user[`skills_${skill}`]);
 	return convertXPtoLVL(xp);
-}
-
-export function calcMaxTripLength(user: User | KlasaUser, activity?: activity_type_enum) {
-	let max = Time.Minute * 30;
-
-	max += patronMaxTripCalc(user);
-
-	switch (activity) {
-		case 'Nightmare':
-		case 'GroupMonsterKilling':
-		case 'MonsterKilling':
-		case 'Wintertodt':
-		case 'Zalcano':
-		case 'BarbarianAssault':
-		case 'AnimatedArmour':
-		case 'Sepulchre':
-		case 'Pickpocket':
-		case 'SoulWars':
-		case 'Cyclops': {
-			const hpLevel = skillLevel(user, SkillsEnum.Hitpoints);
-			const hpPercent = calcWhatPercent(hpLevel - 10, 99 - 10);
-			max += calcPercentOfNum(hpPercent, Time.Minute * 5);
-			break;
-		}
-		case 'Alching': {
-			max *= 2;
-			break;
-		}
-		default: {
-			break;
-		}
-	}
-
-	const sac =
-		user instanceof KlasaUser ? user.settings.get(UserSettings.SacrificedValue) : Number(user.sacrificedValue);
-	const isIronman = user instanceof KlasaUser ? user.isIronman : user.minion_ironman;
-	const sacPercent = Math.min(100, calcWhatPercent(sac, isIronman ? 5_000_000_000 : 10_000_000_000));
-	const perkTier = getUsersPerkTier(user);
-	max += calcPercentOfNum(sacPercent, perkTier >= PerkTier.Four ? Time.Minute * 3 : Time.Minute);
-	return max;
 }
 
 export function getKC(user: KlasaUser | User, id: number) {
